prevent error if Collection.prototype[methodName] doesn't exist
The way tingodb handles findAndModify means it doesn't exist on the prototype which was causing an error to be thrown during `Promise.promisify` call.  
https://github.com/Sanjo/tingodb/blob/master/lib/tcoll.js#L45

I think adding the `_findAndModify200` version to the methodName array will achieve the intended `promisify` result but not sure.  The primary goal of this patch is to prevent the error being thrown.